% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_s.R
\name{arcs.compare}
\alias{arcs.compare}
\title{compare two arcs with no bidirect}
\usage{
arcs.compare(learned, truth)
}
\arguments{
\item{learned}{a data frame of learned arcs}

\item{truth}{a data frame of true arcs}
}
\value{

}
\description{
compare two arcs with no bidirect
}
