% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_s.R
\name{direction.compare}
\alias{direction.compare}
\title{compare two directions}
\usage{
direction.compare(learned, truth)
}
\arguments{
\item{learned}{a data frame of learned arcs}

\item{truth}{a data frame of true arcs}
}
\value{

}
\description{
compare two directions
}
