% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CITL.R
\name{inferTLPs}
\alias{inferTLPs}
\title{Infer time-lagged causality}
\usage{
inferTLPs(
  spliced,
  velocity,
  genes = colnames(spliced),
  n_cluster = 1,
  maxsx = floor(sqrt(ncol(spliced)))
)
}
\arguments{
\item{spliced}{a data frame of mRNA expression data. the colname of it is gene names.}

\item{velocity}{a data frame of velocity data. the colname of it is gene names.}

\item{genes}{the gene name list}

\item{n_cluster}{the number of computational threads}

\item{maxsx}{the maximum allowed size of the conditioning sets used in constructing causal graph}
}
\value{
a summary data frame of time-lagged causal pairs
}
\description{
CITL addresses the limitations of the existing methods for inferring time-lagged causal relationships among genes on single-cell RNA sequencing scRNA-seq data. It adopts the changing information of genes estimated by RNA velocity for causal inference algorithm. Based on the Time-lagged assumption, CITL can infer time-lagged relationships which are validated by published literature. More information and demonstration is in out article (Wei et al., 2021).Before using CITL, annotations of unspliced/spliced reads could be obtained using velocyto CLI or kallisto first.
}
