% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_s.R
\name{rm.bidirect}
\alias{rm.bidirect}
\title{given a arcs(not a graph), return a no bidirect arcs
and a half bidirect arcs}
\usage{
rm.bidirect(arcs_df)

rm.bidirect(arcs_df)
}
\arguments{
\item{arcs_df}{}
}
\value{



}
\description{
given a arcs(not a graph), return a no bidirect arcs
and a half bidirect arcs

given a arcs(not a graph), return a no bidirect arcs
and a half bidirect arcs
}
