% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CITL-data.R
\docType{data}
\name{test_TLPs}
\alias{test_TLPs}
\title{test data for CITL}
\format{
two numeric dataframe and a gene name list.
}
\description{
test data for CITL
}
\examples{
data(test_TLPs);head(test_spliced);head(test_velocity);head(test_gene_names)

}
